+++ backend/src/test/java/develup/api/SolutionApiTest.java
@@ -13,6 +13,7 @@ import java.util.List;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import develup.application.solution.SolutionRequest;
 import develup.application.solution.SolutionResponse;
+import develup.application.solution.StartSolutionRequest;
 import develup.domain.solution.Solution;
 import develup.domain.solution.SolutionSummary;
 import develup.support.data.MemberTestData;
